model/requirements.txt
python-dotenv
transformers[torch]
sentence-transformers
alpaca-py
numpy
pandas
scikit-learn
matplotlib
